hsi-3835 [BE] Incorpora lógica para persistir el identificador de paciente en api salud

+++ back-end/interoperability/src/main/java/net/pladema/cipres/domain/OutpatientConsultationBo.java
@@ -18,7 +18,7 @@ public class OutpatientConsultationBo {
 
 	private BasicDataPatientBo patient;
 
-	private Integer apiPatientId;
+	private Long apiPatientId;
 
 	private String date;
